requirements.txt
gunicorn
Pillow
pandas
ultralytics
flask
flask-cors
torch
torchvision
ultralytics
opencv-python-headless
numpy